[번개] 시큐리티 적용 및 일부 에러 수정 (#104)
* [번개] 버튼 클릭시 클릭 active 켜기

* [번개] 검색조건 완료

* [번개] category-placeId 컬럼 삭제 및 PlaceCategoryMap 엔티티 추가, 엔티티 변경으로 인해 에러나는 jpql 에러수정

* [번개] 번개 게시물 상세 - 댓글 구현 완료 (#64)

* [번개] gatheringCommentTest 생성 및  대댓글 기능 생성 테스트 구현 완료

* [번개] 댓글 작성 기능 구현

* [번개] 답댓글 데이터 저장 완료

* [번개⚡️] 대댓글 정보 DB서버에서 불러오기 완료

* [번개] 답댓글 달때 댓글 다는 사람 뱃지 추가!

* [번개] 댓글에서 멘션없이 답글 달면 nullpointException나는 버그 수정

* [번개] 멘션 댓글 달 시 폼 안에 뱃지 넣기

* [번개] 멘션 이상한 곳에 되는 버그 수정

* [번개] 멘션 기능 구현 완료

* [번개] 수정, 삭제 버튼 생성

* [번개] 댓글 수정 버튼 클릭 시 수정화면폼으로 변경

* [번개] depth1용 댓글수정폼 구현완료

* [번개] 댓글수정-취소 클릭 시 원래 댓글 화면으로 변환

* [번개] 댓글 수정 구현 완료

* [번개] 댓글 삭제 구현완료

* [번개] 오타 수정

* [번개] 답장 버튼 클릭 시 @다정으로 디폴트 멘션 잡히는 버그 수정

* [번개] mention이 null이 되는 버그 수정

* [번개] gatheringtest에 placecategorymap 테스트데이터추가

* [번개] querydsl dependency 추가 및 웹에서 criteria 조건 제대로 읽어오기 성공

* [번개] 카테고리 필터링 기능 구현중

* [번개] 기존 ajax 방식 검색

* [번개] 검색화면 폼 구현완료

* [번개] 검색내용에 작성자, 제목+내용 추가 및 쿼리 개선

* [시큐리티] 유저 중복여부 확인을 이메일 -> 카카오유저아이디로 변경 완료

* [번개] 번개게시글, 댓글에 시큐리티 적용 완료

* [번개] 댓글& 대댓글 작성시 유저닉네임으로 유저를 구분하는 방식에서 아이디로 유저를 구분하는 방식으로 변경(닉네임은 중복이 될 수 있기 때문)

* [번개] 로그인/로그아웃 버튼 생성

* [번개] 게시글상세 &댓글상세 - 수정, 삭제 버튼이 작성자에게만 보이도록 처리

* [번개] 대댓글이 없는 댓글의 경우에는 댓글이 삭제되고, 그 외의 경우에만 댓글내용수정으로 처리한다.

* [번개] 번개화면에 슬라이드바 추가

* [번개] 로그인을 하지 않아도 게시글 상세화면을 볼 수 있도록 설정

* [번개] 로그인 하지 않은 회원에게는 댓글 textarea disabled로 보이게 설정

* [번개] 사이드바 아이콘 클릭 시 나오도록 설정완료

* [번개] 검색버튼 꾸미기 & 메인문구 클릭시 검색 초기화

* [번개] 검색form에서 엔터 누를 시 제출되어버리는 버그 수정

+++ your-anmok/src/main/java/com/lion/youranmok/gathering/controller/GatheringListController.java
@@ -10,10 +10,15 @@ import com.lion.youranmok.gathering.service.GatheringCommentService;
 import com.lion.youranmok.gathering.service.GatheringService;
 import com.lion.youranmok.place.entity.Place;
 import com.lion.youranmok.place.service.PlaceService;
+import com.lion.youranmok.security.dto.MemberContext;
+import com.lion.youranmok.user.entity.User;
+import com.lion.youranmok.user.service.UserService;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -44,9 +49,12 @@ public class GatheringListController {
 
     private final PlaceService placeService;
 
+    private final UserService userService;
+
 
     @GetMapping("/list")
-    public String getList(Model model, @ModelAttribute GatheringListCriteria criteria) {
+    public String getList(@AuthenticationPrincipal MemberContext member, Model model, @ModelAttribute GatheringListCriteria criteria) {
+
 //        System.out.println(criteria.toString());
 
         List<GatheringListDetailDto> gatheringBoardList = gatheringService.listByCriteria(criteria);
@@ -60,7 +68,7 @@ public class GatheringListController {
 
 
     @GetMapping("/{id}")
-    public String getBoardDetail(@PathVariable int id, Model model, CommentForm commentForm) {
+    public String getBoardDetail(@AuthenticationPrincipal MemberContext member, @PathVariable int id, Model model, CommentForm commentForm) {
         GatheringDetailDto gatheringDetailDto = gatheringService.getDetailById(id);
         List<CommentDto> commentList = gatheringCommentService.listByBoardId(id);
         for(CommentDto comment : commentList) {
@@ -75,21 +83,25 @@ public class GatheringListController {
         return "gathering/detail";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/create")
-    public String createBoard(Model model) {
+    public String createBoard(@AuthenticationPrincipal MemberContext member, Model model) {
         return "gathering/create";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create")
-    public String create(Model model, CreateForm createForm) {
-        System.out.println(createForm.toString());
-        GatheringBoard gatheringBoard = gatheringService.create(createForm);
+    public String create(@AuthenticationPrincipal MemberContext member, Model model, CreateForm createForm) {
+        User user = userService.findByUsername(member.getUsername());
+//        System.out.println(createForm.toString());
+        GatheringBoard gatheringBoard = gatheringService.create(user, createForm);
         return "redirect:/gathering/%d".formatted(gatheringBoard.getId());
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/create/search")
     @ResponseBody
-    public ModelAndView searchPlace(@RequestParam String searchKeyword, ModelAndView mav) {
+    public ModelAndView searchPlace(@AuthenticationPrincipal MemberContext member, @RequestParam String searchKeyword, ModelAndView mav) {
 //        System.out.println("searchKeyword:"+searchKeyword);
         List<CreateSearchDto> dataList =  gatheringService.findCreateSearchResultByKeyword(searchKeyword);
 
@@ -101,13 +113,20 @@ public class GatheringListController {
         return mav;
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String getGatheringBoardModify(@PathVariable int id, CreateForm createForm) {
+    public String getGatheringBoardModify(@AuthenticationPrincipal MemberContext memberContext, @PathVariable int id, CreateForm createForm) {
 //        System.out.println(createForm.toString());
         GatheringBoard gatheringBoard = gatheringService.findById(id);
-        Place place = placeService.getPlace(id);
+
+        Place place = placeService.getPlace(gatheringBoard.getPlace().getId());
         if(gatheringBoard == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 게시글이 존재하지 않습니다.");
+        }
+
+        User user = userService.findByUsername(memberContext.getUsername());
+        if(gatheringBoard.getUserId() != user.getId()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "본인이 작성하지 않은 게시글은 수정이 불가능합니다.");
         }
 
         createForm.setGatherCnt(gatheringBoard.getGatherCnt());
@@ -121,16 +140,39 @@ public class GatheringListController {
         return "gathering/modify_form";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{id}")
-    public String doGatheringBoardModify(@PathVariable int id, CreateForm createForm){
-//        System.out.println(createForm.toString());
+    public String doGatheringBoardModify(@AuthenticationPrincipal MemberContext memberContext, @PathVariable int id, CreateForm createForm){
+        GatheringBoard gatheringBoard = gatheringService.findById(id);
+
+        if(gatheringBoard == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 게시글이 존재하지 않습니다.");
+        }
+
+        User user = userService.findByUsername(memberContext.getUsername());
+        if(gatheringBoard.getUserId() != user.getId()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "본인이 작성하지 않은 게시글은 수정이 불가능합니다.");
+        }
         gatheringService.modify(id, createForm);
 
         return "redirect:/gathering/%s".formatted(id);
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String deleteBoard(@PathVariable int id) {
+    public String deleteBoard(@AuthenticationPrincipal MemberContext memberContext, @PathVariable int id) {
+        GatheringBoard gatheringBoard = gatheringService.findById(id);
+
+        if(gatheringBoard == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 게시글이 존재하지 않습니다.");
+        }
+
+        User user = userService.findByUsername(memberContext.getUsername());
+
+        if(gatheringBoard.getUserId() != user.getId()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "본인이 작성하지 않은 게시글은 수정이 불가능합니다.");
+        }
+
         gatheringService.delete(id);
 
         return "redirect:/gathering/list";
